refactor: Rename GitSource to GitException
