+++ Corgibytes.Freshli.Cli.Test/Functionality/GitCommitIdentifierTest.cs
@@ -1,3 +1,4 @@
+using Corgibytes.Freshli.Cli.Exceptions;
 using Corgibytes.Freshli.Cli.Functionality.Git;
 using Corgibytes.Freshli.Cli.Test.Common;
 using Xunit;
